重新整理了common-api下的实体类所属包，避免JPA的entity scan出现问题
