+++ common-api/src/main/java/com/devdynamo/entity/order/OrderItem.java
@@ -1,4 +1,4 @@
-package com.devdynamo.entity;
+package com.devdynamo.entity.order;
 
 import jakarta.persistence.*;
 import lombok.Data;
