+++ common-api/src/main/java/com/devdynamo/entity/payment/Payment.java
@@ -1,4 +1,4 @@
-package com.devdynamo.entity;
+package com.devdynamo.entity.payment;
 
 import com.devdynamo.constant.PaymentStatus;
 import jakarta.persistence.*;
