+++ common-api/src/main/java/com/devdynamo/entity/product/Category.java
@@ -1,4 +1,4 @@
-package com.devdynamo.entity;
+package com.devdynamo.entity.product;
 
 import jakarta.persistence.*;
 import lombok.Data;
