+++ common-api/src/main/java/com/devdynamo/entity/product/Product.java
@@ -1,23 +1,30 @@
-package com.devdynamo.entity;
+package com.devdynamo.entity.product;
 
 import jakarta.persistence.*;
 import lombok.Data;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Data
 @Entity
-@Table(name = "cart_item")
-public class CartItem {
+@Table(name = "product")
+public class Product {
+
+    // TODO 尚不完整
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+    private String name;
+    private String description;
+    private String picture;
+    private Double price;
 
-    @ManyToOne
-    private Cart cartId;
-
-    private Long productId;
-    private Integer quantity;
+    /**
+     * 是不是能自动创建
+     */
+    @ManyToMany(fetch = FetchType.EAGER)
+    private List<Category> categories;
 
     @Column(name = "create_time")
     private LocalDateTime createTime;
@@ -29,6 +36,11 @@ public class CartItem {
     @PrePersist
     public void prePersist() {
         this.createTime = LocalDateTime.now();
+        this.updateTime = LocalDateTime.now();
+    }
+
+    @PreUpdate
+    public void preUpdate() {
         this.updateTime = LocalDateTime.now();
     }
 }
