+++ order-service/src/main/java/com/devdynamo/order/repository/AddressRepository.java
@@ -1,6 +1,6 @@
 package com.devdynamo.order.repository;
 
-import com.devdynamo.entity.Address;
+import com.devdynamo.entity.order.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
