+++ order-service/src/main/java/com/devdynamo/order/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package com.devdynamo.order.repository;
 
-import com.devdynamo.entity.Order;
+import com.devdynamo.entity.order.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
