+++ order-service/src/test/java/com/devdynamo/order/AddressRepositoryTest.java
@@ -1,11 +1,10 @@
 package com.devdynamo.order;
 
-import com.devdynamo.entity.Address;
-import com.devdynamo.entity.Order;
+import com.devdynamo.entity.order.Address;
+import com.devdynamo.entity.order.Order;
 import com.devdynamo.order.repository.AddressRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import java.util.List;
