+++ payment-service/src/main/java/com/devdynamo/payment/repository/PaymentRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.devdynamo.constant.PaymentStatus;
-import com.devdynamo.entity.Payment;
+import com.devdynamo.entity.payment.Payment;
 
 @Repository
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
