+++ product-service/src/main/java/com/devdynamo/product/ProductServiceApplication.java
@@ -1,10 +1,18 @@
 package com.devdynamo.product;
 
-import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.data.elasticsearch.ElasticsearchDataAutoConfiguration;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.boot.autoconfigure.elasticsearch.ElasticsearchRestClientAutoConfiguration;
 
-@SpringBootApplication
+
+@SpringBootApplication(exclude = {
+        ElasticsearchDataAutoConfiguration.class,
+        ElasticsearchRestClientAutoConfiguration.class
+})
+@EntityScan("com.devdynamo.entity.product")
 //@EnableDubbo
 public class ProductServiceApplication {
