+++ product-service/src/main/java/com/devdynamo/product/repository/ProductElasticsearchRepository.java
@@ -1,7 +1,7 @@
-package com.devdynamo.product.repository;
-
-import com.devdynamo.product.entity.elasticsearch.Product;
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
-
-public interface ProductElasticsearchRepository extends ElasticsearchRepository<Product, String> {
-}
+//package com.devdynamo.product.repository;
+//
+//import com.devdynamo.product.entity.elasticsearch.Product;
+//import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
+//
+//public interface ProductElasticsearchRepository extends ElasticsearchRepository<Product, String> {
+//}
