+++ product-service/src/main/java/com/devdynamo/product/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.devdynamo.product.repository;
 
-import com.devdynamo.entity.Product;
+import com.devdynamo.entity.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
