+++ product-service/src/main/resources/application.yaml
@@ -11,9 +11,9 @@ dubbo:
     address: nacos://localhost:8848
     register-mode: instance
 
-spring:
-  elasticsearch:
-    uris: http://localhost:9200
+#spring:
+#  elasticsearch:
+#    uris: http://localhost:9200
 #    username: jjr
 #    password: 12345678
