+++ user-service/src/main/java/com/devdynamo/user/repository/UserRepository.java
@@ -2,7 +2,7 @@ package com.devdynamo.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.devdynamo.entity.User;
+import com.devdynamo.entity.user.User;
 import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
