+++ user-service/src/main/resources/application.properties
@@ -22,10 +22,3 @@ jwt.expiration=604800000
 # Redis配置
 spring.data.redis.host=localhost
 spring.data.redis.port=6379
-
-# Nacos配置
-spring.config.import=optional:nacos:${spring.application.name}.${spring.cloud.nacos.config.file-extension}
-#spring.application.name=user-service
-spring.cloud.nacos.config.file-extension=properties
-spring.cloud.nacos.discovery.enabled=true
-spring.cloud.nacos.config.enabled=true
