--- user-service/src/main/resources/bootstrap.properties
@@ -1,29 +1,0 @@
-# 应用名称
-spring.application.name=user-service
-
-# Nacos认证配置
-#spring.cloud.nacos.username=nacos
-#spring.cloud.nacos.password=nacos
-#spring.cloud.nacos.server-addr=127.0.0.1:8848
-
-# Nacos配置中心
-spring.cloud.nacos.config.enabled=true
-#spring.cloud.nacos.config.server-addr=127.0.0.1:8848
-#spring.cloud.nacos.config.username=${spring.cloud.nacos.username}
-#spring.cloud.nacos.config.password=${spring.cloud.nacos.password}
-spring.cloud.nacos.config.file-extension=properties
-spring.cloud.nacos.config.group=DEFAULT_GROUP
-spring.cloud.nacos.config.namespace=public
-spring.cloud.nacos.config.context-path=/nacos
-
-# Nacos服务发现
-#spring.cloud.nacos.discovery.enabled=true
-#spring.cloud.nacos.discovery.server-addr=127.0.0.1:8848
-#spring.cloud.nacos.discovery.username=${spring.cloud.nacos.username}
-#spring.cloud.nacos.discovery.password=${spring.cloud.nacos.password}
-#spring.cloud.nacos.discovery.context-path=/nacos
-
-# Spring Cloud配置
-spring.cloud.compatibility-verifier.enabled=false
-# 允许覆盖Bean定义
-spring.main.allow-bean-definition-overriding=true
